No need for class attributes.
You use class and object attributes with the same names (__word, __type, __def, __tran). But in fact, these class attributes are not used and not needed.

If I understand correctly, you want to do C++ - style (where in the class we must declare all the attributes, but in Python there is no such need).